+++ segno/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (c) 2016 - 2019 -- Lars Heuer - Semagia <http://www.semagia.com/>.
+# Copyright (c) 2016 - 2020 -- Lars Heuer
 # All rights reserved.
 #
 # License: BSD License
@@ -21,7 +21,7 @@ try:  # pragma: no cover
 except NameError:  # pragma: no cover
     str_type = str
 
-__version__ = '0.3.3'
+__version__ = '0.3.4'
 
 __all__ = ('make', 'make_qr', 'make_micro', 'make_sequence', 'QRCode',
            'QRCodeSequence', 'QRCodeError', 'ErrorLevelError', 'ModeError',
@@ -400,12 +400,12 @@ class QRCode:
         import tempfile
         import webbrowser
         import threading
-        try:  # Python 2
-            from urlparse import urljoin
-            from urllib import pathname2url
-        except ImportError:  # Python 3
+        try:  # Python 3
             from urllib.parse import urljoin
             from urllib.request import pathname2url
+        except ImportError:  # Python 2
+            from urlparse import urljoin
+            from urllib import pathname2url
 
         def delete_file(name):
             time.sleep(delete_after)
@@ -616,6 +616,12 @@ class QRCode:
 
         **Portable Network Graphics (PNG)**
 
+        This writes either a grayscale (maybe with transparency) PNG (color type 0)
+        or a palette-based (maybe with transparency) image (color type 3).
+        If the color / background values are ``None``, white or black, the serializer
+        chooses the more compact grayscale mode, in all other cases a palette-based
+        image is written.
+
         =============    ==============================================================
         Name             Description
         =============    ==============================================================
@@ -636,8 +642,12 @@ class QRCode:
                          that the DPI value is converted into meters (maybe with
                          rounding errors) since PNG does not support the unit
                          "dots per inch".
-        addad            Boolean value (default: True) to (dis-)allow a "Software"
-                         comment indicating that the file was created by Segno.
+        colormap         Optional module type -> color mapping. If provided, the
+                         "color" and "background" arguments are ignored. All undefined
+                         module types will have the default colors (light: white,
+                         dark: black).
+                         See ""color" for valid color values. ``None`` is accepted as
+                         valid color value as well (becomes transparent).
         =============    ==============================================================